Provide the ability to encrypt TPM storage.
Create a generic manager for flash block storage that abstracts the details about how the data is actually stored on flash.  Current implementations provide the ability to store data in plain or cipher text.

Keystores in flash and TPM storage have been updated to use the flash block storage API for flash management.