Merged PR 1472: Generate TCG log in FW
To speed up TCG log generation, construct log in FW. The TCG log is added as another log type, though the log size command will not support the TCG log type.

A further possible optimization is to maintain a ready copy of the log, only updated as needed, so retrieval time is only the time it takes to perform the I2C transactions. This might not be worth the time and effort since the log retrieval time is probably dominated by the time it takes to read out the log over I2C. Speeding up the I2C bus is another possible avenue to explore.

Related work items: #695