Create a placeholder with the minimum definition of log header format 0xCC.
The log header needs to be forward compatible to support consistent parsing across different versions of requester and responder.  While this new header will not be used in current firmware, a placeholder with a minimum definition has been created to allow requesters to be updated with support for this new structure.