Update flash block storage to support minimum required page writes.
Some flash devices require a full page to be written.  With these devices, it is not possible to partially program a page, then add either a header or hash to the beginning or end of the data.  An internal page buffer is required to ensure these disjoint pieces can be aggregated and written together to the flash page.