Merged PR 1757: Fix bug #1029: Possible MCTP buffer overflow
Fix bug #1029: Possible MCTP buffer overflow

Add a check for header->byte_count to be at least sizeof(struct
mctp_protocol_transport_header) to prevent possible buffer overflow if
the value is manipulated.

Add unit tests to test this change, as well as catching a buffer size
that is too small. The buffer size check was already in the code, but
was not being unit tested.

Related work items: #1029