Add a device-side implementation for the OCP Recovery protocol.
The OCP Recovery specification is a work in progress, but this provides an implementation based on its current state.  This is an implementation to enable a device to use the protocol to provide a recovery interface.  There is no implementation for Recovery Agent functionality.

It is not a complete solution as a specific device would need to provide the platform-dependent pieces to fully enable recovery flows, but all the protocol components are implemented such that they do not directly depend on any specific HW.