Create a utility to execute the OCP Recovery protocol.
A utility was created against the OCP Recovery protocol spec to read/write commands.  This can be used for both testing a device implementation as well as execute real-world production workflows, such as image recovery and log retrieval.  It is not tied to any specific implementation, but does provide some hooks for Cerberus-specific handling with log regions.  It also does not provide 100% coverage of the OCP spec, but it covers most cases.  Specifically, polling CMS regions are not currently handled.

There was also an update to the spec, changing the name of PROT_ID to DEVICE_ID.  The device-side implementation of the protocol was updated accordingly.