Significant updates to the OCP recovery interface test script.
The first iteration of this script drove specific tests internally to the script.  This update changes the core script to just be an execution and verification engine.  The actual test cases to execute are defined in JSON files that get passed in.  Each test is generally a sequence of application commands to execute, optionally with information to validate after execution.  This gives the user flexibility to define tests and expectations for different devices.

Not every test is achievable solely through sequences of application commands, so there currently remains a command to completely fill and verify a CMS with random data that is orchestrated by the test script.  When and how this executes is still driven by the JSON definition.  This model can be extended with additional internal commands as they are found to be necessary.

In addition to script changes, the ability to set the SMBus read block size was removed, since the amount of data being sent is not specified by the requestor.