Fix a portability issue with the mock framework.
The mock implementation relies on casting arguments and return values to intptr_t to have consistent handling of both integers and pointers.  On 64-bit machines this works fine.  However, when compiling for 32-bit machines, this means mocks that use 64-bit integers may not be working correctly, even if the tests are passing.  Arguments passed as part of expectations or function calls will get truncated to 32-bits.  Since both the expectation and call are similarly affected, comparisons during validation will pass.  This same limitation applies to return values specified in expectations.  It will not be possible to generate a return value that uses more than the lower 32 bits.

To address this problem, arguments and return values are now cast to int64_t to ensure constant sizing across architectures.  However, this approach causes a pointer-to-int-cast warning on 32-bit machines since the pointers are being cast to a larger container.  The mock has been updated to require differentiation between integer and pointer arguments and return values to ensure proper casting.  All existing tests and mocks have been updated accordingly.

This change only affects test code and no test functionality has been altered.