Enhance the firmware component API to allow for additional header data to be read from memory
For firmware components that have extra header data prepended to them, this header data is necessary for component verification.  In some cases, it is desirable to have this header data already present in memory rather than reading it from flash, particularly when loading the component data into memory.  New APIs have been added to allow pre-loaded header data to be included as part of the firmware component verification.

The generic image header API was also updated to manage the extra header information.