Support API-only initialization for signature verification handlers.
To have parity with static initialization flows, add functions to only initialize the API of ECC and RSA signature verification handlers.  It is now possible to update the instance with a call to init the API followed by a call to init the variable state, which is similar to the static initialization workflow.