Refactor the FreeRTOS platform layer to separate the clock implementation.
It is desirable in some cases to not use the FreeRTOS platform_clock implementation based on the OS tick count.  To support the ability to choose a different platform_clock implementation while still maintain the rest of the FreeRTOS platform API, separate the clock pieces out into a different file and provide a compile-time configuration that will include the alternative implementation through platform.h.  If nothing is defined, the standard tick count clock implementation will be used.

As another small, unrelated change, fix the SPDM command interface to build correctly when CMD_ENABLE_ISSUE_REQUEST is not set.