When trying to restore a recovery, enforce firmware header requirements.
Although the firmware header is only being read from the source image to determine what the resulting recovery revision will be, if the source image fails verification because it lacks the required firmware header, it doesn't make sense to proceed with the recovery restore operation using that source image.  Just leave the recovery image as-is and marked as bad.  The next successfully validated update will come and fix it.