Provide a SPI flash API to force a device reset.
There may be scenarios where a flash device reset is needed without any checks for device state.  The state tracking and other checks that exist in the standard reset call could potentially block the reset in these cases, so provide an API that bypasses all checks and just issues a device reset.