Merged PR 2380: Add host flash reset on host resets
Add the host flash reset on host resets as an add-on to the existing reset flow.  New host processor derived types are added to override soft reset and pass along a flag to indicate host flash reset to host_processor_filtered.  If the reset flash flag is set, a host flash reset will happen whether flash verification happens or not.

Related work items: #1680