Create a new platform API for signaling semaphores from an interrupt handler.
Some platforms (e.g. FreeRTOS) require that different functions be used to signal a semphore from a user task vs. and interrupt handler.  The platform API for semaphores has been updated to accommodate this requirement.  For platforms that don't have such a distinction, both version of the call will be the same and can be statically #defined in this way.