Update SPI filter IRQ handler to support const instances.
The SPI filter IRQ handler API was changed to take a const instance and the existing handlers were updated with static initializers.

Additionally, a new set of common math functions have been added to allow bit checking and manipulation over an arbitrary length array of bytes.  This is useful is scenarios when there are more than 64 bits of flags being used to track state, such as when monitoring SPI opcodes that are blocked by a SPI filter.