Decouple X.509 certificate creation and custom extension management.
Instead of creating and adding custom extensions directly as part of the X.509 implementations, the APIs have been updated to support external management of custom extensions.  This allows the X.509 certificate build process and APIs to remain constant regardless of how many custom extensions are being added to the certificate and what type they are.  This also allows easy updating or customization to the extensions themselves.

This decoupling is the first step in supporting DME extensions in the CSRs, since it is necessary to support CSR with and without DME extensions, as well as DME extensions of varying structures.

The DICE layer 0 handler (RIoT core) has also been updated to support these new APIs.

This is also the first in a series of changes to reorganize some ASN.1/DER handling out of the 'crypto' directory and into a new location.  The goal is also to have ASN.1/DER processing natively supported in the code base, using the RIoT reference code as a starting point.  However, the goal is to deprecate and remove the RIoT code in favor of fully tested code in the common code base.  The base64_riot implementation was updated to support this goal.

OID handling has been updated to remove assumptions about being a valid string.  OIDs can contain zeros, which would get interpreted as a NULL.  Changed all OID handling to bytes with a length.