Create a test utility for running unseal flows with a target device.
There already existed a script to seal data, but unseal needed a target.  This adds a utility that uses the unseal handler that would be present in a target, allowing that workflow to be run in a dev environment.

This just executes the unseal and doesn't include any of the command handling.

This implementation required a null keystore to be created.  The keystore API was opportunistically updated to support const instances and static init.