Update TCG export handling to account for misaligned accesses.
The event size field in the TCG event structure is not word aligned, which causes issues on some architectures.  Update the TCG export process to use a stack variable that is aligned for the event size, while accessing the misalgined field of the packed structure directly.  This allows the compiler to compensate as necessary.