Create a common handler for managing authorization tokens.
Several workflows require authorization tokens, but the construction and requirements around these tokens will vary.  Provide a common handler to securely manage authorization tokens for any purpose.  The only structural requirement is that the nonce and signature be at the end of the token.

This implementation is based off the token handling from authorization_challenge, but provides several enhancements over that flow:
1. Supports signing with any hash algorithm, both for the token and from the authorizing entity.
2. Loads keys on demand to allow for easy write protection of critical security assets.
3. Supports time-based expiration of tokens.
4. Allows arbitrary length data to be prepended to the nonce.
5. Supports arbitrary length nonces.
6. Supports appending additional data to the token before signing by the authorizing entity.  This allows data sourced externally to be bound to the device token and authorized for use by the device.
7. Provides the option for static allocation of the token buffer.

A later change will refactor authorization_challenge to use this new token manager.