Update message structure to allow protocol isolation.
The existing message handling requires that all layers of the command processing stack be aware of the lower layer protocol headers present on the message.  This is fine if the lower layers are always the same, but in scenarios where the transport protocol is different, having this understanding is limiting and complex.

Update the message structure to manage the current location of the message payload, skipping whatever protocol headers are present in the buffer.  This allows command definitions and handlers to be written without needing to know what wrapping protocol was used for the message.

This change represents a placeholder for future use of this feature.  Nothing has changed except the SPDM response MCTP processing, and even that isn't leveraged by the SPDM requester in any way.  Cerberus, MCTP, and SPDM message definitions and handling would need to be updated to make use of the payload pointer in the message structure in those scenarios.