Don't report init errors for bad manifest contents.
Any error handling the contents of a manifest in flash should be ignored during init.  Exposing any of these errors externally would cause system initialization to fail, and since the flash is not automatically erased, this error is persistent.  Instead, treat all errors from manifest, PFM, CFM, or PCD handling as equivalent and just mark the region as invalid.