Create a handler for retrieving SPDM measurement blocks.
This creates an interface to be used by SPDM responders for retrieving measurements from the device.  It maps the SPDM-style requests to calls against the measurement storage, generating a valid SPDM measurement block for the data.

While introducing this handler, there was some refactoring and renaming of existing measurement components:
- Naming of types and macros were updated to better align with SPDM nomenclature.
- All the device discovery components were moved to a separate header file.

This only implements the APIs for getting single measurement blocks.  Implementations for the remaining APIs will follow in a separate commit.