Create a configuration file to use uncrustify for code formatting.
This configuration (or maybe the tool itself) does not generate perfectly formatting, but is a reasonable first step to auto-formatting.

This PR was generated using the latest version of uncrustify (v0.78.1) built from source on Ubuntu 20.04.  It has shown to make much better formatting decisions, in some cases, relative to the default version of 0.68 that gets installed in Ubuntu 20.04 through apt-get.  However, it seems to have trouble and not run on large files (possibly with >10k lines).

There are a few notable issues with the current formatting:
1. There does not seem to be increased indenting with additional scopes in a control statement or function call.
2. Blank line decisions in the presence of #endif don't work great (so that config is current disabled).
3. Aligning spaces are inserted in macro definitions leading up to a \ character.
4. It seems that if the first parameter in a function call needs to be wrapped, the function call itself is wrapped also.  But, the call and arguments are all indented to the same level.
5. There are some cases where extra indent levels are added.
6. New line points for wrapping cases don't always make the best decision for readability.
7. Ternary statements seem to be handled inconsistently.