Create new components for better command handling.
The existing command handling infrastructure is tightly coupled with Cerberus protocol messages over SMBus MCTP.  While it's possible to support other messages, doing some usually means touching existing pieces or carrying unnecessary support for some things.  This change does not alter any of the existing command handling, but it provides the groundwork for future changes that will.

The initial goal of this work was simply to define a generic command handler that could decode multiple different types of messages from within a protocol message.  For example, rather than MCTP having explicit knowledge of supported message types and needing to add a new handler to the instance for each new message type, it would be better to have a single command handler in the MCTP layer that could then decode and route different message types.  Different message types could be added or removed from support by simply changing the single handler provided to the MCTP layer.  The goal was to do this in a way so that it could be used for any protocol that could have a similar use-case (e.g. Cerberus rq bit, SPDM vendor defined messages, PCISIG messages, etc.).

After initially building the handler for multi-message protocols (cmd_interface_multi_handler), some cmd_interface_protocol implementations were created to serve as a reference for introducing new protocols using this framework.  The additional intent here was to build the pieces necessary to immediately decouple specific message type handling from the MCTP layer, enabling easy integration of the SPDM responder code.  Over the course of this work, the level to which pieces are dependent on each other or make assumptions about the protocol stack became apparent.  All of the Cerberus message processing has also not been migrated to the updated cmd_interface_msg structure.  This all means a full switch is not likely possible in the near term, but notes have been added for things that need updating.  Additional refactoring notes have been left for pieces that should be moved or removed (e.g. cmd_interface functions and dependencies), to support cleaner, more generic usage.

Conceptually, this new approach is not that much different from what is there now.  However, the split between protocol handling and message processing is more obvious and independent.  The new structure allows for much better isolation between the layers of the protocol stack, which should provide improved usage flexibility and extensibility.

With this architecture, it should be possible to achieve a fully abstract approach to command handling.  Handling for each individual command could be implemented as a cmd_interface derivation, taking all dependencies needed for execution at init time.  A system would decide which commands to support by instantiating the desired types of command handlers.  These handlers would be registered as a list of handlers to a cmd_interface_multi_handler, which would route based on the command ID (as provided by t...