Refactor hash self-tests to optimize execution.
It's not always necessary to test all the different ways of interfacing with the hash engine, since data management is not part of the algorithm.  For hash self-tests, separate the two different calculation mechanisms into separate self-tests so that both don't always get run for any given algorithm.  For HMAC self-tests, remove the start/update/finish set of HMAC calls, as this is redundant.