Update attestation to request a single certificate at a time from a device.
Requesting the entire certificate chain and building it in dynamically allocated memory is not scalable to arbitrary certificate chain lengths.  The certificate chain processing has been updated to only get one certificate at a time from the device and authenticate it, starting from the root CA.  As each certificate is validated, the trust anchor of the chain is moved.  This ensures there are no more than two certificates in memory any given time.  It does have the side effect of no longer enforcing path length constraints on CA certificates.

This change was applied to both Cerberus and SPDM based attestation.  In addition, most situations will completely rely on the static message buffer for certificate storage, removing dynamic allocation out of the request path.  Overall, this approach may be slightly less performant, especially for SPDM, but is much more efficient in terms of memory usage.