Updates to command authorization handing.
This is the first part of several steps to support impactless vs. impactful detection in the FW update process.

The existing command authorization handling is not easily scalable to support different numbers and types of authorization handlers.  Every new authenticated command would need an explicit API added to the interface for handling and the init call would get extended.  This would also require a more complicated integration if certain platforms had authenticated commands that didn't exist on others.  The command authorization has been updated to remove these inefficiencies by providing a single API that would work for any arbitrary command.  The supported commands and identifiers for these commands is configured per-platform at init time.  The existing APIs for specific commands have been left for now, but may be removed in the future.

In addition, the command authorization handler was updated to support static and const instances.  This change had the ripple effect of requiring updates to the authorization contexts as well.  These were equally updated to support static and const instances.

Authorization using a challenge token was even further updated, since the current implementation is limited to using SHA-256 hashes for token signing and verifying.  To provide flexibility to support any combination of keys and hashes, the challenge handler was updated to use the auth_token type for token management.  This makes challenge authorization not much more than an adaptor between the authorization and auth_token APIs.