Create a message transport for handling MCTP VDM encapsulation.
This will handle any type of vendor defined, and it may or may not be a necessary layer of the protocol stack depending on how the vendor protocol is constructed.

As part of this update, commonalities between all these intermediate layers of the protocol stack has been factored out into a new type that each of the MCTP message transports inherits.