Create the framework for executing arbitrary authorized commands.
Update to the background command handler to support executing arbitrary commands that have been securely authorized.  This is the first step of this refactoring to provide the hooks for this handling.  A subsequent update will provide execution contexts for the current commands and update the command processing to use these new interfaces.