Extend the FW update observer to provide a notification during the prepare stage.
There exists a mechanism for updates to be blocked before applying them to boot flash.  However, this doesn't happen until after the staging flash has been programmed and verified.  To provide earlier feedback of the failure, and to do it without changing any flash contents, extend the update observer interface to have the same mechanism in the 'prepare' stage of the update flow.