Handler to detect and manage impactful firmware updates.
This is a set of helper modules that will be used by the firmware updater to determine if a particular update will be impactful, and if so, if that update has been authorized.

The part that actually determines if an update is impactful is abstracted behind and interface.  Multiple different contexts can be checked for different situations that could result in an impactful update.  Each check also has the ability to block authorized overrides and not allow the update to be accepted by the device.  Some examples of possible impactful checks are compatibility checks with the new firmware image or certain SPDM session states.