Merged PR 3869: Fix Get Message type support to set completion_code properly
Current testcases are not able to identify this issue as they
are doing memset to zero to response data, Hence change in testcase
is also required.

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix to ensure proper setting of the completion code in the MCTP GET Message Type support response.

#### PR Summary
This pull request addresses an issue where the MCTP GET Message Type support response was not setting the completion code correctly, leading to bad completion codes.
- `core/mctp/mctp_control_protocol_commands.c`: Added setting of `completion_code` to `MCTP_CONTROL_PROTOCOL_SUCCESS` in the response header.
- `core/testing/mctp/mctp_control_protocol_commands_test.c`: Updated test to set `completion_code` to `0x80` for validation.

Related work items: #4183