Rename AES API to explicitly call out GCM operation.
There is no longer an expectation for a single AES API to support different modes.  Instead, there are two APIs:  one for GCM and one for XTS.  Rename the old API which supports GCM to make this obvious.

This touches several files and adds some "new" ones, but these are just renames of the older files.