Merged PR 3874: Initial changes for variable TPM buffer size
TPM currently has a fixed block size of 512 bytes.
For newer platforms, which support larger block sizes, a flexible block size is needed.
As part of this change, the tpm_init API is modified to allocate TPM with a variable block size.
Unit testing was done on Overlake-3

Cerberus changes are in PR:
https://rackmanager.visualstudio.com/Cerberus/_git/Cerberus/pullrequest/3924

Related work items: #2024