Update TcbInfo extension building to accept arbitrary length SVN values.
Some platforms have support for more SVN values than can fit into 32 bits.  To handle these cases, the TcbInfo extension builders need to be more flexible.  Rather than a fixed, 32-bit integer for SVN, the builder has been updated to accept an arbitrary length array of SVN bytes.  It requires that the caller format the integer value correctly in the array, but it allows for any length SVN to be encoded.