Merged PR 3996: Add an API in TPM to get segment-storage-size
With flexible block size of TPM storage, an API is needed to get the block size.
This PR adds the same.
Associated mock tests are added as well.

----
#### AI description  (iteration 1)
#### PR Classification
New feature

#### PR Summary
This pull request adds an API to retrieve the segment storage size in TPM, along with associated tests.
- Added `tpm_get_segment_storage_size` function in `core/tpm/tpm.c` and declared it in `core/tpm/tpm.h`.
- Implemented tests for the new API in `core/testing/tpm/tpm_test.c`.

Related work items: #4253