Merged PR 3989: Added error code for Host firmware image activation failure
This PR is to add error code required for firmware image activation failures this error code facilitates a text message at which stage of activation it is failed and the corresponding error code details in the debug log.

----
#### AI description  (iteration 1)
#### PR Classification
New feature

#### PR Summary
This pull request adds a new error code for activation failure to enhance error logging and recovery mechanisms.
- `core/host_fw/host_logging.h`: Added `HOST_LOGGING_ACTIVATION_FAILURE` to log Host image activation failures.

Related work items: #3762