Update the DICE key manager to support additional CSRs that can be exported.
In some cases, the device may have other CSRs (or public key information) that it wants to export.  This data can be exported using the Export CSR command with a non-zero index.  The DICE key manager has been updated to manage these additional CSRs that may be present.

As part of this update the key manager was also updated to support static initialization and const instances.  Several comments have also been updated to reference 'DICE' instead of 'RIoT'.