Support mbedTLS 2.28.x
For the most part, 2.28 vs. 2.17 are the same, as no APIs have changed.  There were only a couple differences.
1. ECC P-521 DER encoding includes the leading zero on the point bitstream while 2.17 excluded this.
2. v2.28.9 doesn't include the NULL entry for the signature algorithm when generating X.509 certificates.

Both of these differences generate valid output, but the unit tests needed to account for the different expectations.

The folder name for mbedTLS when checking out was also changed to decouple it from a specific version, leaving that up to the integration to decide what to use.