Update testing engine macro definitions to support implementations that require state.
As part of migration to const/static crypto instances, some implementations will require a variable 'state' structure to be defined and passed during init.  The current macros used to abstract the specific implementation for a crypto engine used during unit tests do not account for this dependency.  The macros for all engines have been updated to support engines that do and do not require a 'state' structure.

This change has means using the macros slightly differently and needing to use the macros in more scenarios than previously was necessary.  This has resulted in extensive changes to the unit test code, though the changes are nothing more than syntax.  Functionality hasn't been altered.

The mbedTLS hash implementation was updated to require a state structure as part of verification of the new macro definitions.  Full support for const/static initialization will come later.