Update to support building on Ubuntu 24.04
A couple things are different on the versions of Ubuntu that caused issues with the current code:
- gcc13
- OpenSSL 3.x

For gcc13, this primarily meant updating mbedTLS, which was done previously.  But there were also a few spots in the rest of the code that triggered build warnings on the new compiler.

For OpenSSL 3.x, there was significant rework to the ECC and RSA implementations.  Conceptually, the change was simply that everything needed to interface with the library through the high-level EVP_PKEY abstraction rather than dealing with lower-level EC_KEY and RSA types.  However, there were also a few scenarios that didn't have a common method of handling on both v1.x and 3.x, so these parts are compile-time switched using a macro that looks at the version being used.

This change does not remove support for older versions (e.g. Ubuntu-20.04).  It's compatible with both environments.