Fix deadlock issue when the complete SPDM certificate chain is not retrieved successfully.
Both hash engines are used by the requester for long-running hash operations:
- The secondary hash engine was always the transcript hash, which is started with the first SPDM message and not finished until the last response with a signature is received.
- The primary hash engine is used to calculate the certificate chain digest, which spans multiple SPDM messages now that each cert in the chain is individually retrieved from the device.

The primary hash engine management when used with retrieved certificate was not correct.  If an intermediate or leaf certificate failed to be retrieved from the device, the active hash on the cert chain would not be canceled.  This would leave the hash active indefinitely, blocking any future hash operations with that hash engine.  If this hash engine was something with a mutex, like hash_engine_thread_safe, this would result in deadlock.

Since there are long-running hash operations executed in the SPDM context, the hash engines used by the SPDM requester should not be used by other system components.  To ensure that attestation requester workflows don't interfere with other system functions, the requester needs two hash engines dedicated for its use that are not shared with anything else.