Provide a single function to call for all signature verification flows.
Per FIPS, signature verification involves both hash and verify of the message data.  Wrap these steps in a single function to provide a single algorithm implementation that can be self-tested and used wherever signature verification is needed.  The verification itself using the signature_verification abstraction, so it works for both ECDSA and RSASSA.

A set of self-tests have also been added for this function.