Update the deterministic ECDSA implementation to align with the RFC for P-521 keys.
For k values that are not even multiples of bytes, the output needs to be right shifted to generate the RFC-compliant value.  The previous implementation just masked the first bits out.

Basically, the RFC requires that the last generated bits get dropped while the previous shortcut implementation dropped the first generated bits.

Also update the deterministic ECDSA function signatures to better align with typical calling structure (i.e. the associated instance is the first argument).