Update the hash API with a function to query the active hash algorithm.
In scenarios where an active hash context is passed to another module, it would be useful to be able to query the hash engine for the type of hash being run.  This is especially needed in cases where the digest is retrieved from the hash engine.  Rather than needing to have an argument passed along with the hash engine, update the API for the hash engine to provide this information directly.  This also ensures that the information is coming directly from the piece that is in control of it.