Remove hash algorithm argument from ECDSA APIs.
In any scenario where ECDSA is being executed against a hash that has already been started externally, do not take the hash algorithm as a parameter to the API.  Instead, query the hash engine directly to determine the digest length for the active hash.

When ECDSA is hashing the full massage, the algorithm still needs to be provided to indicate how ECDSA should generate the digest.