Provide a handler for verifying device firmware with a PFM.
It may be beneficial in some scenarios to leverage a PFM for verifying and measuring a firmware image on flash that will be used by the device.  This is different than the host firmware verification flow in that the PFM is expected to be part of and managed with the device's firmware.

This does not provide any facility to authenticate the image as it's being loaded, so it's not a complete solution for booting securely.  This approach is primarily meant to provide similar PFM measurements for internal RoTs where firmware loading my not all happen within the RoT.  For example, Linux scenarios where the RoT loads U-Boot directly but delegates responsibilities for kernel loading and authentication to U-Boot might want to pre-check the Linux image against a PFM and get a measurement.

A few other changes were made to existing pieces to enable this type of workflow:
1. Host firmware verification functions that were being reused with this new verification flow were updated to decouple from spi_flash and instead work with the generic flash interface.
2. The new workflow measures the same information for the PFM as existing host firmware handing.  A new function was added to manifest_pcr to expose the common measurement workflows to this context.
3. Handling PFM measurement data for TCG log construction also was updated to leverage common flows, with some updates to manifest_manager to enable this reuse.