Add clarifying comment about mbedTLS GCM decrypt API usage.
MbedTLS decrypt documentation indicates that the input and output buffers can't be the same.  However, there is no evidence of this requirement in the current code.  In fact, there is evidence that this has not been a requirement of the API since 2013.  Adding documentation to acknowledge the documented usage of the API and why we aren't adhering to that.