Provide an HKDF implementation that is FIPS compliant.
FIPS requires that the HKDF algorithm implementation execute both the HKDF-Extract and HKDF-Expand operations in the same context.  When deriving multiple keys, it needs to be ensured that all keys come from the same Extract step.  A new interface has been defined with a general implementation to provide the state management and APIs necessary to derive keys using HKDF.

KDF selt-tests have been updated to test the new algorithm implementation.

As part of this change, the HMAC implementation was updated to support zero-length keys.  While not that useful from a security perspective, a zero-length HMAC key is not invalid.  HKDF leverages this capability in cases where there is no salt provided.