Merged PR 4827: Add riot common instance static initialization with given API
There are scenarios where riot core common instance is sufficient for all DICE Layer 0 handling except for the need to override some APIs.  This PR adds a macro that can be utilized for this purpose.

----
#### AI description  (iteration 1)
#### PR Classification
New feature

#### PR Summary
This pull request adds a new static initialization method for the Riot core common instance with a given API.
- `core/riot/riot_core_common_static.h`: Added `riot_core_common_static_init` macro for initializing a static instance with a specified API and updated documentation for the internal initialization function.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #5081